Add xvfb to avoid a bug in chromedriver.
browser.serValue() causes the following error:
Failed: unknown error: an X display is required for keycode conversions,
consider using Xvfb

This is a chromedriver's bug that is fixed in version 2.31.
This commit is a workaround until the 2.31 will be released for Alpine
Linux.

ref. https://github.com/SeleniumHQ/docker-selenium/issues/429